new server mod that separates chatlogs out into their own file
great for parsing or auditing. logs all chat messages WITH the channel indicator, as well as player index and timestamp, and if applicable, alliance index or sector the message was sent in. includes some rudimentary command parsing and tab-delimiting of most of the fields so you can write your regexps or just have your output look somewhat reasonably uniform